Act on a per-buffer basis instead of per-window
I feel like people like buffers more. Basing the highlighting on windows was
useful in order to allow the same file (well, buffer) open in multiple windows
to be highlighted in some and not in others; at the same time, it wouldn't allow
to have different states in the same window (I can think of at least a couple of
cases, involving NERDTree or vim-vinegar, where that isn't desirable).